# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-grade, high-performance, distributed-ready multi-agent orchestration engine",
        "title": "Goclaw API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/goclaw/goclaw",
            "email": "support@goclaw.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/workflows": {
            "get": {
                "description": "List all workflows with optional filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "List workflows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of workflows",
                        "schema": {
                            "$ref": "#/definitions/models.WorkflowListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Submit a new workflow for execution with tasks and dependencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Submit a new workflow",
                "parameters": [
                    {
                        "description": "Workflow definition",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WorkflowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Workflow created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.WorkflowResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/workflows/{id}": {
            "get": {
                "description": "Get the current status and details of a specific workflow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Get workflow status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workflow status",
                        "schema": {
                            "$ref": "#/definitions/models.WorkflowStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid workflow ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Workflow not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/workflows/{id}/cancel": {
            "post": {
                "description": "Cancel a running or pending workflow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Cancel a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workflow cancelled successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid workflow ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Workflow cannot be cancelled",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/workflows/{id}/tasks/{tid}/result": {
            "get": {
                "description": "Get the result of a specific task within a workflow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Get task result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task result",
                        "schema": {
                            "$ref": "#/definitions/models.TaskResultResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid workflow ID or task ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task result not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is alive and running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service is unhealthy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Check if the service is ready to accept requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness check",
                "responses": {
                    "200": {
                        "description": "Service is ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "503": {
                        "description": "Service is not ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Get detailed status information about the service and engine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Detailed status",
                "responses": {
                    "200": {
                        "description": "Detailed status information",
                        "schema": {
                            "$ref": "#/definitions/engine.EngineStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "engine.EngineStatus": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.TaskDefinition": {
            "type": "object",
            "required": [
                "id",
                "name",
                "type"
            ],
            "properties": {
                "config": {
                    "description": "Config holds task-specific configuration.",
                    "type": "object",
                    "additionalProperties": true
                },
                "depends_on": {
                    "description": "DependsOn lists task IDs that must complete before this task.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "task-0"
                    ]
                },
                "id": {
                    "description": "ID is the unique task identifier within the workflow.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "task-1"
                },
                "name": {
                    "description": "Name is the task name.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Fetch data from API"
                },
                "retries": {
                    "description": "Retries is the number of retry attempts on failure.",
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 0,
                    "example": 3
                },
                "timeout": {
                    "description": "Timeout is the maximum execution time in seconds.",
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 1,
                    "example": 300
                },
                "type": {
                    "description": "Type is the task type (e.g., \"http\", \"script\", \"function\").",
                    "type": "string",
                    "enum": [
                        "http",
                        "script",
                        "function"
                    ],
                    "example": "http"
                }
            }
        },
        "models.TaskResultResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt is when the task completed.",
                    "type": "string"
                },
                "error": {
                    "description": "Error holds error information if the task failed.",
                    "type": "string"
                },
                "result": {
                    "description": "Result holds the task result data."
                },
                "status": {
                    "description": "Status is the task status.",
                    "type": "string"
                },
                "task_id": {
                    "description": "TaskID is the task identifier.",
                    "type": "string"
                },
                "workflow_id": {
                    "description": "WorkflowID is the workflow identifier.",
                    "type": "string"
                }
            }
        },
        "models.TaskStatus": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt is when the task completed.",
                    "type": "string"
                },
                "error": {
                    "description": "Error holds error information if the task failed.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the task identifier.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the task name.",
                    "type": "string"
                },
                "result": {
                    "description": "Result holds the task result data."
                },
                "started_at": {
                    "description": "StartedAt is when the task started.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the current task status.",
                    "type": "string"
                }
            }
        },
        "models.WorkflowListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit is the maximum number of results returned.",
                    "type": "integer"
                },
                "offset": {
                    "description": "Offset is the starting position in the result set.",
                    "type": "integer"
                },
                "total": {
                    "description": "Total is the total number of workflows matching the filter.",
                    "type": "integer"
                },
                "workflows": {
                    "description": "Workflows is the list of workflow summaries.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WorkflowSummary"
                    }
                }
            }
        },
        "models.WorkflowRequest": {
            "type": "object",
            "required": [
                "name",
                "tasks"
            ],
            "properties": {
                "description": {
                    "description": "Description is an optional workflow description.",
                    "type": "string",
                    "maxLength": 500,
                    "example": "Process customer data and generate reports"
                },
                "metadata": {
                    "description": "Metadata holds optional key-value pairs.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "environment": "production",
                        "team": "data-engineering"
                    }
                },
                "name": {
                    "description": "Name is the workflow name.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "data-processing-workflow"
                },
                "tasks": {
                    "description": "Tasks is the list of tasks in the workflow.",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.TaskDefinition"
                    }
                }
            }
        },
        "models.WorkflowResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt is the workflow creation timestamp.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique workflow identifier.",
                    "type": "string"
                },
                "message": {
                    "description": "Message provides additional information.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the workflow name.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the current workflow status.",
                    "type": "string"
                }
            }
        },
        "models.WorkflowStatusResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt is when the workflow completed.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt is the workflow creation timestamp.",
                    "type": "string"
                },
                "error": {
                    "description": "Error holds error information if the workflow failed.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the workflow identifier.",
                    "type": "string"
                },
                "metadata": {
                    "description": "Metadata holds workflow metadata.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the workflow name.",
                    "type": "string"
                },
                "started_at": {
                    "description": "StartedAt is when the workflow started executing.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the current workflow status.",
                    "type": "string"
                },
                "tasks": {
                    "description": "Tasks is the list of task statuses.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TaskStatus"
                    }
                }
            }
        },
        "models.WorkflowSummary": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt is when the workflow completed.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt is the workflow creation timestamp.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the workflow identifier.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the workflow name.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the current workflow status.",
                    "type": "string"
                },
                "task_count": {
                    "description": "TaskCount is the total number of tasks.",
                    "type": "integer"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        }
    }
}